Move indicator parsing code to general util
