+++ utils/index.ts
@@ -1,3 +1,4 @@
 export * from './apollo';
 export * from './data';
+export * from './select';
 export * from './strings';
